fix(backend): Refactor Controller and routes for users, orders, and products.
